+++ src/structure/mod.rs
@@ -1,8 +1,8 @@
 //! データ構造っぽいもの
 pub mod binary_trie;
 pub mod cumsum_2d;
+pub mod disjoint_set_union;
 pub mod disjoint_sparse_table;
-pub mod dsu;
 pub mod fenwick_tree;
 pub mod segment_tree;
 pub mod skew_heap;
